requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pydantic==1.8.2
python-dotenv==0.19.1
SQLAlchemy==1.3.8
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
